requirements.txt
google-cloud-speech==1.3.2
google-cloud-storage==1.27.0